Fixed migrations and added JS

+++ src/migrations/m201031_132510_remove_unused_site_id.php
@@ -15,19 +15,15 @@ class m201031_132510_remove_unused_site_id extends Migration
      */
     public function safeUp()
     {
-        // try {
-        //     $this->dropForeignKey("matrixfieldpreview_previewrecord_siteId_fk", "{{%matrixfieldpreview_blocktypes_config}}");
-        // } catch (Exception $e) {
-        // }
+        $table = Craft::$app->db->schema->getTableSchema('{{%matrixfieldpreview_blocktypes_config}}');
+        if (isset($table->columns['siteId'])) {
+            $this->dropColumn("{{%matrixfieldpreview_blocktypes_config}}", "siteId");
+        }
 
-        // try {
-        //     $this->dropForeignKey("matrixfieldpreview_blocktypes_config_siteId_fk", "{{%matrixfieldpreview_blocktypes_config}}");
-        // } catch (Exception $e) {
-        // }
-
-        // $this->dropForeignKey("matrixfieldpreview_fields_config_siteId_fk", "{{%matrixfieldpreview_fields_config}}");
-        $this->dropColumn("{{%matrixfieldpreview_blocktypes_config}}", "siteId");
-        $this->dropColumn("{{%matrixfieldpreview_fields_config}}", "siteId");
+        $table = Craft::$app->db->schema->getTableSchema('{{%matrixfieldpreview_fields_config}}');
+        if (isset($table->columns['siteId'])) {
+            $this->dropColumn("{{%matrixfieldpreview_fields_config}}", "siteId");
+        }
     }
 
     /**
